--- src/aavs_uv/calibration/__init__.py
@@ -1,3 +1,0 @@
-from .simple_cal import simple_stefcal
-from .simple_sim import simulate_visibilities
-from .stefcal import stefcal
