--- src/aavs_uv/postx/__init__.py
@@ -1,3 +1,0 @@
-from .ant_array import RadioArray
-from .allsky_viewer import AllSkyViewer
-from .sky_model import generate_skycat
