+++ tests/test_converter.py
@@ -14,11 +14,11 @@ def test_converter():
                     os.remove(f"test.{fmt}")
 
             # Create command-line args
-            cmd = ["-n", "aavs2", 
-                "-o", fmt, 
-                "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+            cmd = ["-n", "aavs2",
+                "-o", fmt,
+                "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
                 f"test.{fmt}"]
-            
+
             # Run the script
             args = parse_args(cmd)
             run(cmd)
@@ -32,9 +32,9 @@ def test_converter():
 
 def test_custom_config():
     try:
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-        "-o", "sdp", 
-        "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+        "-o", "sdp",
+        "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
         "test.sdp"]
         run(cmd)
     finally:
@@ -45,17 +45,17 @@ def test_phase_to_sun():
     try:
         cmd = [
         "-o", "uvfits", "-s", "-n", "aavs3",
-        "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+        "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
         "test.uvfits"]
         run(cmd)
     finally:
         if os.path.exists("test.uvfits"):
-            os.remove("test.uvfits")    
+            os.remove("test.uvfits")
 
 def test_errors():
         # Create command-line args
-        cmd = ["-c", "carmen/santiago", 
-            "-o", "svg", 
+        cmd = ["-c", "carmen/santiago",
+            "-o", "svg",
             "input.hdf5",
             "output.sdp"]
         # Run the script
@@ -64,17 +64,17 @@ def test_errors():
 
 def test_batch():
     try:
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-b", 
-               "-o", "uvx", 
-               "../example-data/aavs2_2x500ms",
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-b",
+               "-o", "uvx",
+               "./test-data/aavs2_2x500ms",
                "test-batch-data"]
         run(cmd)
 
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-B", 
-               "-o", "uvx", 
-               "test-data/", # Note -B MEGABATCH test 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-B",
+               "-o", "uvx",
+               "test-data/", # Note -B MEGABATCH test
                "test-batch-data"]
         run(cmd)
     finally:
@@ -83,19 +83,19 @@ def test_batch():
 
 def test_batch_zip():
     try:
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-b", 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-b",
                "-z",
-               "-o", "ms", 
-               "../example-data/aavs2_2x500ms",
+               "-o", "ms",
+               "./test-data/aavs2_2x500ms",
                "test-batch-data"]
         run(cmd)
 
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-B", 
-               "-z", 
-               "-o", "miriad", 
-               "test-data/", # Note -B MEGABATCH test 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-B",
+               "-z",
+               "-o", "miriad",
+               "test-data/", # Note -B MEGABATCH test
                "test-batch-data2"]
         run(cmd)
 
@@ -108,11 +108,11 @@ def test_batch_zip():
 
 def test_batch_multi():
     try:
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-b", 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-b",
                "-z",
-               "-o", "ms,miriad,uvx", 
-               "../example-data/aavs2_2x500ms",
+               "-o", "ms,miriad,uvx",
+               "./test-data/aavs2_2x500ms",
                "test-batch-data"]
         run(cmd)
 
@@ -124,16 +124,16 @@ def test_batch_multi():
 
 def test_context():
     try:
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-i", "test-data/context.yml", 
-               "-o", "uvx", 
-               "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-i", "test-data/context.yml",
+               "-o", "uvx",
+               "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
                "test.uvx5"]
         run(cmd)
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-i", "test-data/context.yml", 
-               "-o", "sdp", 
-               "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-i", "test-data/context.yml",
+               "-o", "sdp",
+               "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
                "test.sdp"]
         run(cmd)
         uv = read_uvx("test.uvx5")
@@ -148,37 +148,37 @@ def test_context():
 def test_parallel():
     try:
         # One worker
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-i", "test-data/context.yml", 
-               "-o", "sdp", 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-i", "test-data/context.yml",
+               "-o", "sdp",
                "-w", "2",
-               "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+               "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
                "test.sdp"]
         run(cmd)
         # Change number of workers
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-i", "test-data/context.yml", 
-               "-o", "sdp", 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-i", "test-data/context.yml",
+               "-o", "sdp",
                "-w", "8",
-               "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+               "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
                "test.sdp"]
         run(cmd)
         # Use dask instead
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-i", "test-data/context.yml", 
-               "-o", "sdp", 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-i", "test-data/context.yml",
+               "-o", "sdp",
                "-w", "8",
                "-p", "dask",
-               "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+               "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
                "test.sdp"]
         run(cmd)
         # Verbose
-        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'), 
-               "-i", "test-data/context.yml", 
-               "-o", "sdp", 
+        cmd = ["-c", get_resource_path('config/aavs3/uv_config.yaml'),
+               "-i", "test-data/context.yml",
+               "-o", "sdp",
                "-w", "1",
                "-v",
-               "../example-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
+               "./test-data/aavs2_2x500ms/correlation_burst_204_20230927_35116_0.hdf5",
                "test.sdp"]
         run(cmd)
 
@@ -190,10 +190,11 @@ def test_parallel():
 
 if __name__ == "__main__":
     test_batch_multi()
-    #test_batch()
-    #test_context()
-    #test_phase_to_sun()
-    #test_custom_config()
-    #test_errors()
-    #test_converter()
-    #test_batch_zip()
+    test_batch()
+    test_context()
+    test_phase_to_sun()
+    test_custom_config()
+    test_errors()
+    test_converter()
+    test_batch_zip()
+    test_parallel()
